Update RIR to GAP
