--- django_project/core/views/__init__.py
@@ -1,2 +1,0 @@
-__author__ = 'Irwan Fathurrahman <meomancer@gmail.com>'
-__date__ = '13/10/21'
