--- django_project/gap_dashboard/forms/__init__.py
@@ -1,2 +1,0 @@
-__author__ = 'Irwan Fathurrahman <meomancer@gmail.com>'
-__date__ = '04/11/21'
