+++ django_project/gap_data/admin/dashboard.py
@@ -2,27 +2,25 @@
 from django.contrib import admin
 
 from gap_data.models.dashboard import (
-    Dashboard, Plugin
+    Dashboard, Widget
 )
 
 
-class PluginInline(admin.StackedInline):
-    """Plugin inline."""
+class WidgetInline(admin.StackedInline):
+    """Widget inline."""
 
-    model = Plugin
+    model = Widget
     extra = 0
 
 
 class DashboardAdmin(admin.ModelAdmin):
     """Dashboard admin."""
 
-    list_display = (
-        'slug', 'name',
-    )
+    list_display = ('slug', 'name',)
     filter_horizontal = (
-        'reference_layers', 'basemap_layers', 'context_layers'
+        'basemap_layers', 'context_layers'
     )
-    inlines = (PluginInline,)
+    inlines = (WidgetInline,)
     prepopulated_fields = {'slug': ('name',)}
