+++ django_project/gap_data/admin/indicator.py
@@ -7,17 +7,18 @@ from gap_data.models.indicator import (
 )
 
 
+class IndicatorExtraValueRuleInline(admin.TabularInline):
+    """IndicatorExtraValue inline."""
+
+    model = IndicatorExtraValue
+    extra = 0
+
+
 class IndicatorValueAdmin(admin.ModelAdmin):
     """IndicatorValue admin."""
 
-    class IndicatorExtraValueRuleInline(admin.TabularInline):
-        """IndicatorExtraValue inline."""
-
-        model = IndicatorExtraValue
-        extra = 0
-
-    list_display = ('indicator', 'date', 'geometry', 'value')
-    list_filter = ('indicator', 'date', 'geometry')
+    list_display = ('indicator', 'date', 'geom_identifier', 'value')
+    list_filter = ('indicator', 'date', 'geom_identifier')
     search_fields = ('indicator',)
     inlines = (IndicatorExtraValueRuleInline,)
 
@@ -28,19 +29,17 @@ class IndicatorFrequencyAdmin(admin.ModelAdmin):
     list_display = ('name', 'frequency')
 
 
+class IndicatorRuleInline(admin.TabularInline):
+    """IndicatorRule inline."""
+
+    model = IndicatorRule
+    extra = 0
+
+
 class IndicatorAdmin(admin.ModelAdmin):
     """Indicator admin."""
 
-    class IndicatorRuleInline(admin.TabularInline):
-        """IndicatorRule inline."""
-
-        model = IndicatorRule
-        extra = 0
-
-    list_display = (
-        'name', 'group', 'frequency',
-        'geometry_reporting_level')
-    filter_horizontal = ('geometry_reporting_units',)
+    list_display = ('name', 'group', 'frequency', 'reporting_level')
     list_filter = ('group',)
     inlines = (IndicatorRuleInline,)
