+++ django_project/gap_data/models/__init__.py
@@ -3,3 +3,4 @@ from .context_layer import *
 from .dashboard import *
 from .indicator import *
 from .link import *
+from .reference_layer import *
