+++ django_project/gap_data/models/dashboard/__init__.py
@@ -1,2 +1,2 @@
 from .dashboard import *
-from .plugin import *
+from .widget import *
