+++ django_project/gap_data/models/indicator/indicator_attributes.py
@@ -26,8 +26,7 @@ class IndicatorFrequency(AbstractTerm):
 
 
 class IndicatorGroup(AbstractTerm):
-    """The group of indicator.
-    """
+    """The group of indicator. """
 
     # dashboard link
     dashboard_link = models.CharField(
