+++ django_project/gap_data/models/indicator/indicator_value.py
@@ -2,7 +2,6 @@
 from django.contrib.gis.db import models
 from django.utils.translation import ugettext_lazy as _
 
-from gap_data.models.reference_layer import Geometry
 from gap_data.models.indicator.indicator import Indicator
 
 
@@ -16,14 +15,17 @@ class IndicatorValue(models.Model):
         _('Date'),
         help_text=_('The date of the value harvested.')
     )
-    geometry = models.ForeignKey(
-        Geometry, on_delete=models.SET_NULL,
+    # TODO:
+    #  When recreate all migrations from zero
+    #  Default it removed
+    geom_identifier = models.CharField(
+        max_length=256,
         null=True, blank=True
     )
     value = models.FloatField()
 
     class Meta:  # noqa: D106
-        unique_together = ('indicator', 'date', 'geometry')
+        unique_together = ('indicator', 'date', 'geom_identifier')
         ordering = ('-date',)
 
 
@@ -51,38 +53,3 @@ class IndicatorExtraValue(models.Model):
 
     def __str__(self):
         return f'{self.name}'
-
-
-class IndicatorValueExtraDetailRow(models.Model):
-    """A group name of the extra indicator value.
-
-    It contains extra value in dictionary that saved in
-    IndicatorValueExtraDetailColumn.
-    """
-
-    indicator_value = models.ForeignKey(
-        IndicatorValue, on_delete=models.CASCADE
-    )
-
-
-class IndicatorValueExtraDetailColumn(models.Model):
-    """Additional data for Indicator value data in column index."""
-
-    row = models.ForeignKey(
-        IndicatorValueExtraDetailRow, on_delete=models.CASCADE
-    )
-    name = models.CharField(
-        max_length=100,
-        help_text=_(
-            "The name of column"
-        )
-    )
-    value = models.TextField(
-        null=True, default=True,
-        help_text=_(
-            "The value of cell"
-        )
-    )
-
-    class Meta:  # noqa: D106
-        unique_together = ('row', 'name')
