+++ django_project/gap_data/serializer/dashboard.py
@@ -2,20 +2,18 @@
 
 from rest_framework import serializers
 
-from gap_data.models.dashboard import (
-    Plugin
-)
+from gap_data.models.dashboard import Widget
 
 
-class PluginSerializer(serializers.ModelSerializer):
-    """Serializer for Plugin."""
+class WidgetSerializer(serializers.ModelSerializer):
+    """Serializer for Widget."""
 
     layer_id = serializers.SerializerMethodField()
 
-    def get_layer_id(self, obj: Plugin):
-        """Return layer id of plugin."""
+    def get_layer_id(self, obj: Widget):
+        """Return layer id of Widget."""
         return obj.layer_id
 
     class Meta:  # noqa: D106
-        model = Plugin
+        model = Widget
         exclude = ('reference_layer', 'context_layer')
